Add the log of ring0 calls from Throttlestop
This is the log of all ring0 calls by Throttlestop, from which I inferred the information in this guide. 
The log was created by replacing the winring0.dll from Throttlestop, with a library that first prints the call to a log file and then proceeds to call the real winring0.dll. Credit to Jaka Bac for this shim library.